django/pyproject.toml
[project]
name = "restaurant-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Kaloyan Borisov",email = "kaborisov@tu-sofia.bg"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "djangorestframework (>=3.16.0,<4.0.0)",
    "pillow (>=11.2.1,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
